requirements.txt
dbt-duckdb
dbt-core
metricflow